autodev_generated/requirements.txt
################################################################################
# 核心框架
fastapi==0.110.2
uvicorn[standard]==0.29.0
# ORM & DB
sqlmodel==0.0.16
aiosqlite==0.20.0
# 安全
argon2-cffi==23.1.0
PyJWT[crypto]==2.8.0
cryptography==42.0.7
# 校验 & 配置
pydantic[email]==2.7.1
pydantic-settings>=2.0.0,<3   # 修复：缺失依赖
# 测试
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
# 代码质量
black==24.4.2
ruff==0.4.3